# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arabian Startups Ecosystem API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: number;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  sourceUrl: string;
  sourceName: string;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  country: string;
  tags?: string[];
  isFeatured: boolean;
  viewCount: number;
  publishedAt: string;
  createdAt?: string;
}

export interface ArticleList {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  articleCount: number;
}

export interface CountryStat {
  country: string;
  code: string;
  flag: string;
  articleCount: number;
}

export interface DailyDigest {
  id: number;
  date: string;
  summary: string;
  highlights: string[];
  articleCount: number;
  articles?: Article[];
}

export interface DigestList {
  digests: DailyDigest[];
  total: number;
  page: number;
  limit: number;
}

export type EcosystemStatsArticlesByCategoryItem = {
  category: string;
  count: number;
};

export interface EcosystemStats {
  totalArticles: number;
  totalFundingMentions: number;
  articlesByCountry: CountryStat[];
  articlesByCategory: EcosystemStatsArticlesByCategoryItem[];
  todayCount: number;
  weekCount: number;
}

export interface NewsletterSubscription {
  email: string;
  /** @nullable */
  name?: string | null;
  countries?: string[];
}

export interface SubscriptionResult {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListArticlesParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
country?: string | null;
/**
 * @nullable
 */
search?: string | null;
/**
 * @nullable
 */
date?: string | null;
};

export type GetTrendingArticlesParams = {
limit?: number;
};

export type ListDigestsParams = {
page?: number;
limit?: number;
};