# lib/api-zod/src/generated/types/articleList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arabian Startups Ecosystem API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from './article';

export interface ArticleList {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
}