# lib/api-zod/src/generated/types/countryStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arabian Startups Ecosystem API
 * OpenAPI spec version: 0.1.0
 */

export interface CountryStat {
  country: string;
  code: string;
  flag: string;
  articleCount: number;
}