# lib/api-zod/src/generated/types/dailyDigest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arabian Startups Ecosystem API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from './article';

export interface DailyDigest {
  id: number;
  date: string;
  summary: string;
  highlights: string[];
  articleCount: number;
  articles?: Article[];
}