# lib/api-zod/src/generated/types/digestList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arabian Startups Ecosystem API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyDigest } from './dailyDigest';

export interface DigestList {
  digests: DailyDigest[];
  total: number;
  page: number;
  limit: number;
}