# lib/api-zod/src/generated/types/newsletterSubscription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arabian Startups Ecosystem API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscription {
  email: string;
  /** @nullable */
  name?: string | null;
  countries?: string[];
}